Adding history to app.

+++ app/src/main/java/com/ichinweze/flickpick/data/firestore/FirestoreDataClasses.kt
@@ -16,5 +16,5 @@ data class SelectedMovieDetails(
     val movieTitle: String? = null,
     val posterPath: String? = null,
     val releaseDate: String? = null,
-    val userRating: Float? = null
+    val userRating: Int? = null
 )
